поправки до секціі header
