+++ src/index.js
@@ -1,4 +1,5 @@
 import './sass/main.scss';
-// import Pagination from './js/pagination';
+import './js/get-countries';  
+import Pagination from './js/pagination';
 
-const API_KEY = 'Iks6oDIpGdxCIBqWeGHShYrO2fcgxZEd';
+
