+++ src/partials/header.html
@@ -1,7 +1,7 @@
-<header class="header header-logo">
-    <div class="title-form">
-        <h1 class="title">FIND BEST EVENTS AROUND THE WORLD</h1>
-    </div>
+<header class="header"> <!--header-logo-->
+    <!-- <div class="title-form"> -->
+        <h1 class="title">FIND BEST EVENTS <br> AROUND THE WORLD</h1>
+    <!-- </div> -->
     <form class="search-form" id="search-form">
         <div class="input">
             <input class="input-search"
@@ -9,8 +9,8 @@
                 name="search" 
                 autocomplete="off" 
                 autofocus
-                pattern="^[a-zA-Z\s]+$"
-                placeholder="Start searching" />
+                pattern="^[a-zA-Z\s]+$"  
+                placeholder="Start searching" /> <!--в pattern треба додати кирилицю-->
             <svg class="input-icon">
                 <use href="./images/svg/sprits.svg#icon-search"></use>
             </svg>
