+++ src/sass/components/_footer.scss
@@ -1,11 +1,11 @@
 .footer-section {
-  position: fixed;
+  // position: fixed;
+  left: 0;
+  right: 0;
+  bottom: 20px;
   display: flex;
   flex-direction: column;
   justify-content: center;
-  bottom: 20px;
-  left: 0;
-  right: 0;
   color: var(--blurry-text-color);
 
   @media screen and (max-width: 767px) {
@@ -37,6 +37,17 @@
   animation: animateScale 1700ms infinite cubic-bezier(0.4, 0, 0.2, 1);
 }
 
+.footer-section:hover,
+.footer-section:focus {
+  color: var(--light-text-color);
+  & .heart-icon {
+    animation: animateScale 500ms infinite cubic-bezier(0.4, 0, 0.2, 1);
+  }
+}
+.developers {
+  padding-left: 4px;
+}
+
 @keyframes animateScale {
   0% {
     transform: scale(1);
@@ -50,15 +61,4 @@
   100% {
     transform: scale(1);
   }
-}
-
-.footer-section:hover,
-.footer-section:focus {
-  color: var(--light-text-color);
-  & .heart-icon {
-    animation: animateScale 500ms infinite cubic-bezier(0.4, 0, 0.2, 1);
-  }
-}
-.developers {
-  padding-left: 4px;
 }
